CLS-13: Simplify: Move CommandTemplateLibrary to the dedecated project directory
